requirements.txt
fastapi[standard]
duckduckgo-search
google-api-python-client
openai
python_dotenv